+++ hello_riscv/clint.h
@@ -65,12 +65,12 @@ int clint_ipi_send(size_t hart_id);
 int clint_ipi_clear(size_t hart_id);
 int clint_ipi_register(clint_ipi_cb_t callback, void *user_data);
 
-/* CLINT timer init*/
+/* Clint timer init*/
 int clint_timer_init(void);
 int clint_timer_start(uint64_t delayms, int one_shot);
 int clint_timer_register(clint_timer_cb_t callback, void *user_data);
-int clint_timer_mdelay(uintptr_t ms);
-uintptr_t clint_timer_current_tick(void);
+int clint_timer_mdelay(uint64_t ms);
+uint64_t clint_timer_current_tick(void);
 
 #ifdef __cplusplus
 }
